Refactoring HAJSU żeby SIĘ ZGADZAŁ czyli orderów Collect, Deposit, Production, oraz Resource.value i Army.resource (powrót na inty)

+++ Assets/Scripts/Army.cs
@@ -13,7 +13,7 @@ namespace MechWars
 
         public List<Unit> Units { get; private set; }
 
-        public float resources;
+        public int resources;
 
         public Texture hpBarMain;
         public Texture hpBarSide;
